resources/stage1_requirements.txt
torch
torchvision
pandas
python-magic
lightgbm
lief
tqdm
